requirements.txt
requests
imagehash
opencv-python-headless
Pillow
numpy
torch
